trying to spit files into ui elements and utils, trying to get the correct context when calling addDOMWidget in an alias object
